Update Python bootstrapping details for FreeBSD
##### SUMMARY

When setting up a new FreeBSD host, it is necessary to bootstrap the pkg(8) package manager first. In practice, it's just another yes/no prompt, which requires user interaction. In order to bootstrap pkg(8) automatically, the user must set the ASSUME_ALWAYS_YES environment variable (the `--yes` flag does not work with bootstrapping by design). Let's document that. This also addresses the confusion reported in #15191.

##### ISSUE TYPE
- Docs Pull Request

+label: docsite_pr